# Fix performance regression due to very frequent skin updating